Improve clientside architecture, routing and authenthication using JWT
Change the clientside architecture to the following: The main component
after the app component will be the navbar, which is essential to routing,
The site component is removed but its functionality is covered by the login
component. Login HTTP POST is done by the auth service. Logout is done
directly by the navbar component. This layout has the advantage of the
auth service being re-used due to injectability into navbar and login
components.

Additionally the JWT is stored as cookie on client and used to restrict
access to the 'myfavorites' and 'imageupload' routes via clientside
route-guards (see auth.guard.ts and routing-module). The navbars naming
of the routes was adjusted to our project.